conformance/bindingTesting/queues/cl_commandQueue_enqueueNDRangeKernel.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueNDRangeKernel conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCLCommandQueue enqueueNDRangeKernel test case.");
    debug("( WebCLKernel kernel, CLuint workDim, sequence<CLuint>? globalWorkOffset, sequence<CLuint> globalWorkSize, sequence<CLuint>? localWorkSize, optional sequence<WebCLEvent>? eventWaitList = null, optional WebCLEvent? event = null)");

    var SIZE = 1024;
    var count = SIZE;
    var data = new Float32Array(SIZE);

    var wtu = WebCLTestUtils;
    var webCLPlatform = wtu.getPlatform();
    var webCLDevices = wtu.getDevices(webCLPlatform);
    var webCLContext = wtu.createContext();
    data = wtu.generateRandomFloat(data, count);

    var kernelSource = wtu.readKernel("../../../resources/kernels/square.cl");
    var webCLProgram = wtu.createProgram(webCLContext, kernelSource);
    wtu.build(webCLProgram, webCLDevices);
    var webCLKernel = wtu.createKernel(webCLProgram, "square");

    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);

    var workGroupSize = webCLKernel.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);

    var globalWorkSize = [2, workGroupSize/2];
    var localWorkSize = [2, 1];
    var globalWorkOffset = [1, 1];
    var workDim = globalWorkSize.length;

    input = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    output = wtu.createBuffer(webCLContext, webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    wtu.enqueueWriteBuffer(webCLCommandQueue, input, true, 0, BUFFER_SIZE, data);
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize);", "INVALID_KERNEL_ARGS");

    wtu.setArg(webCLKernel, 0, input);
    wtu.setArg(webCLKernel, 1, output);
    wtu.setArg(webCLKernel, 2, new Uint32Array([count]));

    shouldBeUndefined("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize);");
    shouldBeUndefined("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, [], globalWorkSize, localWorkSize);");
    shouldBeUndefined("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, []);");
    shouldBeUndefined("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, [], globalWorkSize, []);");

    // Testing for Exceptions & invalid values.

    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, 4, globalWorkOffset, globalWorkSize, localWorkSize);", "INVALID_WORK_DIMENSION");

    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, [10, 10, 10], globalWorkSize, localWorkSize);", "INVALID_GLOBAL_OFFSET");
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, [Math.pow(2,32), 1], globalWorkSize, localWorkSize);", "INVALID_GLOBAL_OFFSET");

    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, [], localWorkSize);", "INVALID_GLOBAL_WORK_SIZE");
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, [10], localWorkSize);", "INVALID_GLOBAL_WORK_SIZE");
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, [Math.pow(2,32), 2], localWorkSize);", "INVALID_GLOBAL_WORK_SIZE");

    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, [10, 10, 10]);", "INVALID_WORK_GROUP_SIZE");
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, [300, 400], [4, 3]);", "INVALID_WORK_GROUP_SIZE");

    var deviceMaxWorkGroupSize = webCLDevices[0].getInfo(webcl.DEVICE_MAX_WORK_GROUP_SIZE);
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, [deviceMaxWorkGroupSize, 2]);", "INVALID_WORK_GROUP_SIZE");

    var deviceMaxWorkItemSize = webCLDevices[0].getInfo(webcl.DEVICE_MAX_WORK_ITEM_SIZES);
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, [(deviceMaxWorkItemSize + 10), 2]);", "INVALID_WORK_ITEM_SIZE");

    var webCLdeviceCPU, webCLdeviceGPU, webCLContextALL, webCLKernelCPU, webCLProgramALL, webCLCommandQueueGPU;
    try {
        webCLdeviceCPU = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_CPU)[0];
        webCLdeviceGPU = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_GPU)[0];
        webCLContextALL = wtu.createContext(webCLPlatform, [webCLdeviceCPU, webCLdeviceGPU]);

        webCLProgramALL = wtu.createProgram(webCLContextALL, kernelSource);
        wtu.build(webCLProgramALL, [webCLdeviceCPU]);
        webCLKernelCPU = wtu.createKernel(webCLProgramALL, "square");

        webCLCommandQueueGPU = wtu.createCommandQueue(webCLContext, webCLdeviceGPU);
        shouldThrowExceptionName("webCLCommandQueueGPU.enqueueNDRangeKernel(webCLKernelCPU, workDim, [], globalWorkSize, []);", "INVALID_PROGRAM_EXECUTABLE");
    } catch (e) {
        testFailed("WebCLCommandQueue :: enqueueNDRangeKernel() test for INVALID_PROGRAM_EXECUTABLE threw : " + e.message);
    }

    // FIXME :: need to add test case to test for INVALID_IMAGE_SIZE & MEM_OBJECT_ALLOCATION_FAILURE

    // Test cases for kernel with __attribute [reqd_work_group_size].
    kernelSource = wtu.readKernel("../../../resources/kernels/copy_attributes.cl");
    webCLProgram = wtu.createProgram(webCLContext, kernelSource);
    wtu.build(webCLProgram, webCLDevices);
    webCLKernel = wtu.createKernel(webCLProgram, "copy");

    data = wtu.generateRandomFloat(data, count);
    input = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    output = wtu.createBuffer(webCLContext, webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    wtu.enqueueWriteBuffer(webCLCommandQueue, input, true, 0, BUFFER_SIZE, data);
    wtu.setArg(webCLKernel, 0, input);
    wtu.setArg(webCLKernel, 1, output);

    shouldBeUndefined("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, 3, [], [30, 40, 10], [10, 10, 10]);");
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, [300, 400], [1, 1]);", "INVALID_WORK_GROUP_SIZE");

    var webCLContextNew = wtu.createContext();
    var webCLCommandQueueNew = wtu.createCommandQueue(webCLContextNew);
    var kernelSource = wtu.readKernel("../../../resources/kernels/square.cl");
    var webCLProgramNew = wtu.createProgram(webCLContextNew, kernelSource);
    wtu.build(webCLProgramNew, webCLDevices);
    var webCLKernelNew = wtu.createKernel(webCLProgramNew, "square");
    input = wtu.createBuffer(webCLContextNew, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    output = wtu.createBuffer(webCLContextNew, webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    data = wtu.generateRandomFloat(data, count);
    wtu.enqueueWriteBuffer(webCLCommandQueueNew, input, true, 0, BUFFER_SIZE, data);
    wtu.setArg(webCLKernelNew, 0, input);
    wtu.setArg(webCLKernelNew, 1, output);
    wtu.setArg(webCLKernelNew, 2, new Uint32Array([count]));
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernelNew, workDim, globalWorkOffset, globalWorkSize, localWorkSize);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize);", "INVALID_CONTEXT");

    var webCLEvent = wtu.createEvent(webCLContext);
    var webCLEventNew = wtu.createEvent(webCLContextNew);
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize, [webCLEventNew], webCLEvent);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize, [webCLEvent], webCLEventNew);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize, [], webCLEventNew);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize, [], webCLEvent);", "INVALID_CONTEXT");

    var releasedEvent = wtu.createEvent(webCLContext);
    wtu.release(releasedEvent);
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(webCLKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize, [releasedEvent], webCLEvent);", "INVALID_EVENT_WAIT_LIST");

    var releasedKernel = webCLKernel;
    wtu.release(webCLKernel);
    shouldThrowExceptionName("webCLCommandQueue.enqueueNDRangeKernel(releasedKernel, workDim, globalWorkOffset, globalWorkSize, localWorkSize);", "INVALID_KERNEL");

} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
